# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resonance API Documentation",
        "description": "API documentation for Resonance",
        "contact": {
            "email": "casier07@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://resonance-be.ddev.site",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get a specific event",
                "operationId": "6f743f71f7a88dbe57be1d9c9ebf47cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "operationId": "fb130dcf1ab4d7abbd3a173741df1f8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tomorrowland 2024 - The Story of Planaxis"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-07-19"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-07-28"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The world's biggest electronic dance music festival, featuring the best DJs from around the globe. This year's theme: The Story of Planaxis.",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "concert",
                                            "festival",
                                            "tour",
                                            "clubnight",
                                            "other"
                                        ],
                                        "example": "festival"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://resonance-be.ddev.site/storage/events/tomorrowland-2024-planaxis.jpg",
                                        "nullable": true
                                    },
                                    "source": {
                                        "type": "string",
                                        "enum": [
                                            "manual",
                                            "api"
                                        ],
                                        "example": "manual"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending_approval",
                                            "verified",
                                            "rejected"
                                        ],
                                        "example": "verified"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "operationId": "b8c8f4b41a32532f5e9f7146e511bc61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event deleted successfully"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Tomorrowland 2024"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-19"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-28"
                    },
                    "description": {
                        "type": "string",
                        "example": "The world's biggest electronic dance music festival, featuring the best DJs from around the globe.",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "concert",
                            "festival",
                            "tour",
                            "clubnight",
                            "other"
                        ],
                        "example": "festival"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://resonance-be.ddev.site/storage/events/tomorrowland-2024.jpg",
                        "nullable": true
                    },
                    "source": {
                        "type": "string",
                        "example": "manual"
                    },
                    "status": {
                        "type": "string",
                        "example": "verified"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}